chore: [DHIS2-11862] move orgunit hook
