+++ src/core_modules/capture-core/dataQueries/index.js
@@ -1,0 +1,1 @@
+export { useOrganisationUnit } from './useOrganisationUnit';
